requirements.txt
# Automatically generated by https://github.com/damnever/pigar.

albumentations==1.3.1
matplotlib==3.5.1
MedPy==0.4.0
numpy==1.21.6
opencv-python==4.5.4.58
pandas==1.3.4
Pillow==8.4.0
scikit-image==0.19.2
scipy==1.7.3
SimpleITK==2.2.1
torch==1.7.1
torchvision==0.8.2
tqdm==4.64.0
